% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abesspca.R
\name{abesspca}
\alias{abesspca}
\title{Adaptive best subset selection for principal component analysis}
\usage{
abesspca(
  x,
  type = c("predictor", "gram"),
  cor = FALSE,
  support.size = NULL,
  support.num = NULL,
  c.max = 2,
  lambda = 0,
  always.include = NULL,
  group.index = NULL,
  splicing.type = 1,
  max.splicing.iter = 20,
  warm.start = TRUE,
  ...
)
}
\arguments{
\item{x}{A matrix object. It can be either the predictor matrix where each row is an observation and each column is 
a predictor or the sample covariance/correlation matrix.}

\item{type}{If \code{type = "predictor"}, \code{x} is considered as the predictor matrix. 
If \code{type = "gram"}, \code{x} is considered as a sample covariance or correlation matrix.}

\item{support.size}{An integer vector representing the alternative support sizes. 
Only used for \code{tune.path = "sequence"}. Default is \code{0:min(n, round(n/(log(log(n))log(p))))}.}

\item{c.max}{an integer splicing size. Default is: \code{c.max = 2}.}

\item{lambda}{A single lambda value for regularized best subset selection. Default is 0.}

\item{always.include}{An integer vector containing the indexes of variables that should always be included in the model.}

\item{group.index}{A vector of integers indicating the which group each variable is in.
For variables in the same group, they should be located in adjacent columns of \code{x}
and their corresponding index in \code{group.index} should be the same.
Denote the first group as \code{1}, the second \code{2}, etc.
If you do not fit a model with a group structure,
please set \code{group.index = NULL} (the default).}

\item{max.splicing.iter}{The maximum number of performing splicing algorithm. 
In most of the case, only a few times of splicing iteration can guarantee the convergence. 
Default is \code{max.splicing.iter = 20}.}

\item{warm.start}{Whether to use the last solution as a warm start. Default is \code{warm.start = TRUE}.}

\item{...}{further arguments to be passed to or from methods.}
}
\value{
A S3 \code{abesspca} class object, which is a \code{list} with the following components:
\item{loadings}{A \eqn{p}-by-\code{length(support.size)} loading matrix of sparse principal components (PC), 
where each row is a variable and each column is a support size;}
\item{ev}{A vector with size \code{length(support.size)}. It records the explained variance at each support size.}
\item{var.all}{Total variance of the predictors.}
\item{call}{The original call to \code{abess}.}
}
\description{
Adaptive best subset selection for principal component analysis
}
\examples{
library(abess)

## predictor matrix input:
head(USArrests)  
pca_fit <- abesspca(USArrests)
str(pca_fit)

## covariance matrix input:
pca_fit <- abesspca(stats::cov(USArrests), type = "gram")
str(pca_fit)

## robust covariance matrix input:
rob_cov <- MASS::cov.rob(USArrests)[["cov"]]
rob_cov <- (rob_cov + t(rob_cov)) / 2
pca_fit <- abesspca(rob_cov, type = "gram")
str(pca_fit)

}
\author{
Jin Zhu, Junxian Zhu, Ruihuang Liu, Xueqin Wang
}
