% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.abess.R
\name{coef.abess}
\alias{coef.abess}
\title{Provides estimated coefficients from a fitted "abess" object.}
\usage{
\method{coef}{abess}(object, support.size = NULL, sparse = TRUE, ...)
}
\arguments{
\item{object}{A "\code{abess}" project.}

\item{support.size}{An integer vector specify 
the coefficient fitted at given \code{support.size}. 
If \code{support.size = NULL}, then all coefficients would be returned. 
Default: \code{support.size = NULL}.}

\item{sparse}{A logical value, specifying whether the coefficients should be
presented as sparse matrix or not. Default: \code{sparse = TRUE}.}

\item{...}{Other arguments.}
}
\description{
This function provides estimated
coefficients from a fitted "\code{abess}" object.
}
\seealso{
\code{\link{abess}}, \code{\link{print.abess}}.
}
