% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.abess.R
\name{deviance.abess}
\alias{deviance.abess}
\title{Extract the deviance from a "abess" object.}
\usage{
\method{deviance}{abess}(object, type = c("standard", "gic", "ebic", "bic", "aic"), ...)
}
\arguments{
\item{object}{A "\code{abess}" object.}

\item{\dots}{additional arguments}
}
\description{
Similar to other deviance methods, 
which returns deviance from a fitted "\code{abess}" object.
}
\seealso{
\code{\link{abess}}.
}
