% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.graph.R
\name{find_path}
\alias{find_path}
\title{A recursive function used to find the path that: 
(i) go through all node via the pre-specified edges;
(ii) the path have no loop.}
\usage{
find_path(i, sub_path, adj, n_node)
}
\arguments{
\item{i}{}

\item{sub_path}{}

\item{adj}{}

\item{n_node}{}
}
\description{
A recursive function used to find the path that: 
(i) go through all node via the pre-specified edges;
(ii) the path have no loop.
}
\examples{
index <- find_path(3, c(3), adj, n_node)
length(unique(index)) == n_node
alpha_adj_index <- cbind(index[-n_node], index[-1])
result <- c()
for (i in 1:nrow(alpha_adj_index)) {
  result <- c(result, adj[alpha_adj_index[i, 1], alpha_adj_index[i, 2]])
}
all(result == 1)

}
