% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.graph.R
\name{generate.data.bmn}
\alias{generate.data.bmn}
\title{Generate a dataset from binary markov network}
\usage{
generate.data.bmn(n, p, type = 1, seed, beta = NULL)
}
\arguments{
\item{n}{}

\item{p}{}

\item{type}{}

\item{seed}{}

\item{beta}{}
}
\value{

}
\description{
Generate a dataset from binary markov network
}
