% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{generate.data}
\alias{generate.data}
\title{Generate simulated data}
\usage{
generate.data(
  n,
  p,
  support.size = NULL,
  rho = 0,
  family = c("gaussian", "binomial"),
  beta = NULL,
  cortype = 1,
  snr = 10,
  seed = 1
)
}
\arguments{
\item{n}{The number of observations.}

\item{p}{The number of predictors of interest.}

\item{support.size}{The number of nonzero coefficients in the underlying regression
model. Can be omitted if \code{beta} is supplied.}

\item{rho}{A parameter used to characterize the pairwise correlation in
predictors. Default is \code{0}.}

\item{family}{The distribution of the simulated data. \code{"gaussian"} for
gaussian data.\code{"binomial"} for binary data.}

\item{beta}{The coefficient values in the underlying regression model. 
If it is supplied, \code{support.size} would be omitted.}

\item{cortype}{The correlation structure. 
\code{cortype = 1} denotes the independence structure, 
where the covariance matrix has \eqn{(i,j)} entry equals \eqn{I(i \neq j)}.
\code{cortype = 2} denotes the exponential structure,
where the covariance matrix has \eqn{(i,j)} entry equals \eqn{rho^{|i-j|}}.
code{cortype = 3} denotes the constant structure, 
where the non-diagnoal entries of covariance 
matrix are \eqn{rho} and diagnoal entries are 1.}

\item{snr}{A numerical value controlling the signal-to-noise ratio (SNR). The SNR is defined as
as the variance of \eqn{x\beta} divided
by the variance of a gaussian noise: \eqn{\frac{Var(x\beta)}{\sigma^2}}.
The gaussian noise \eqn{\epsilon} is set with mean 0 and variance.
The noise is added to the linear predictor \eqn{\eta} = \eqn{x\beta}. Default is \code{snr = 10}.
This option is invalid for \code{cortype = 3}.}

\item{seed}{seed to be used in generating the random numbers.}
}
\value{
A \code{list} object comprising:
\item{x}{Design matrix of predictors.} 
\item{y}{Response variable.}
\item{beta}{The coefficients used in the underlying regression model.}
}
\description{
Generate data for simulations under the generalized linear model.
}
\details{
We generate an \eqn{n \times p} random Gaussian matrix
\eqn{X} with mean 0 and a covariance matrix with an exponential structure
or a constant structure. For the exponential structure, the covariance matrix
has \eqn{(i,j)} entry equals \eqn{rho^{|i-j|}}. For the constant structure,
the \eqn{(i,j)} entry of the covariance matrix is \eqn{rho} for every \eqn{i
\neq j} and 1 elsewhere. For the moving average structure,  For the design matrix \eqn{X},
we first generate an \eqn{n \times p} random Gaussian matrix \eqn{\bar{X}}
whose entries are i.i.d. \eqn{\sim N(0,1)} and then normalize its columns
to the \eqn{\sqrt n} length. Then the design matrix \eqn{X} is generated with
\eqn{X_j = \bar{X}_j + \rho(\bar{X}_{j+1}+\bar{X}_{j-1})} for \eqn{j=2,\dots,p-1}.

For \code{family = "gaussian"} , the data model is \deqn{Y = X \beta +
\epsilon.}
The underlying regression coefficient \eqn{\beta} has uniform distribution [m, 100m], \eqn{m=5 \sqrt{2log(p)/n}.}

For \code{family= "binomial"}, the data model is \deqn{Prob(Y = 1) = \exp(X
\beta + \epsilon)/(1 + \exp(X \beta + \epsilon)).}
The underlying regression coefficient \eqn{\beta} has uniform distribution [2m, 10m], \eqn{m = 5\sigma \sqrt{2log(p)/n}.}

In the above models, \eqn{\epsilon \sim N(0,
\sigma^2 ),} where \eqn{\sigma^2} is determined by the \code{snr}.
}
\examples{

# Generate simulated data
n <- 200
p <- 20
support.size <- 5
dataset <- generate.data(n, p, support.size)
str(dataset)

}
\author{
Jin Zhu
}
