% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.spc.matrix.R
\name{generate.spc.matrix}
\alias{generate.spc.matrix}
\title{Generate matrix}
\usage{
generate.spc.matrix(
  n,
  p,
  support.size = 3,
  snr = 20,
  sigma = NULL,
  sparse.loading = NULL,
  seed = 1
)
}
\arguments{
\item{n}{The number of observations.}

\item{p}{The number of predictors of interest.}

\item{support.size}{A integer specify the number of non-zero entries in the first column of loading matrix.}

\item{snr}{A numerical value controlling the signal-to-noise ratio (SNR).
A larger SNR implies the identification of sparse matrix is much easier.}

\item{sigma}{A numerical vector specify the standard deviation of each columns.
Default \code{sigma = NULL} implies it is determined by \code{snr}.}

\item{seed}{random seed. Default: \code{seed = 1}.}
}
\value{
A \code{list} object comprising:
\item{x}{An \eqn{n}-by-\eqn{p} matrix.}
\item{coef}{The sparse loading matrix.}
\item{coef}{A vector recording the number of non-zero entries.}
}
\description{
Generate simulated matrix that its principal component are
sparse linear combination of its columns.
}
\details{
The methods for generating the matrix is detailedly described in the APPENDIX A. Data generation section in Schipper et al (2021).
}
\references{
Model selection techniques for sparse weight-based principal component analysis. de Schipper, Niek C and Van Deun, Katrijn. Journal of Chemometrics. 2021.
}
