% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.abesspca.R
\name{plot.abesspca}
\alias{plot.abesspca}
\title{Creat plot from a fitted "\code{abesspca}" object}
\usage{
\method{plot}{abesspca}(x, type = c("loadings", "variance"), label = FALSE, ...)
}
\arguments{
\item{x}{A "\code{abesspca}" object.}

\item{type}{The type of terms to be plot in the y-axis. 
One of \code{"loadings"} and \code{"variance"}. 
Default is \code{"loadings"}.}

\item{label}{A logical value. 
If \code{label = TRUE} (the default), 
label the curves with variable sequence numbers.}

\item{...}{Other graphical parameters to plot}
}
\value{
No return value, called for side effects.
}
\description{
Produces a coefficient/deviance/tuning-value plot 
for a fitted "\code{abesspca}" object.
}
\examples{
library(abess)
head(USArrests)  
pca_fit <- abesspca(USArrests)
plot(pca_fit)
plot(pca_fit, "variance")

}
\seealso{
\code{\link{print.abesspca}}, 
\code{\link{loadings.abess}}, 
\code{\link{plot.abessca}}.
}
