% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.abess.R
\name{predict.abess}
\alias{predict.abess}
\title{make predictions from a "abess" object.}
\usage{
\method{predict}{abess}(object, newx, type = c("link", "response"), support.size = NULL, ...)
}
\arguments{
\item{object}{Output from the \code{abess} function.}

\item{newx}{New data used for prediction. If omitted, the fitted linear predictors are used.}

\item{type}{\code{type = "link"} gives the linear predictors for \code{"binomial"},
\code{"poisson"} or \code{"cox"} models; for \code{"gaussian"} models it gives the
fitted values. \code{type = "response"} gives the fitted probabilities for
\code{"binomial"}, fitted mean for \code{"poisson"} and the fitted relative-risk for
\code{"cox"}; for \code{"gaussian"}, \code{type = "response"} is equivalent to \code{type = "link"}}

\item{...}{Additional arguments affecting the predictions produced.}
}
\value{
The object returned depends on the types of family.
}
\description{
Returns predictions from a fitted
"\code{abess}" object.
}
\examples{

}
\seealso{
\code{\link{abess}}.
}
