% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.abess.R
\name{print.abess}
\alias{print.abess}
\title{print method for a "abess" object}
\usage{
\method{print}{abess}(x, digits = max(5, getOption("digits") - 5), ...)
}
\arguments{
\item{x}{A "\code{abess}" object.}

\item{digits}{Minimum number of significant digits to be used.}

\item{...}{additional print arguments}
}
\value{
print a \code{data.frame} with three columns:
\item{support.size}{Support size of model} 
\item{dev}{deviance of model}
\item{tune.type}{The tuning value of the certain tuning type}
}
\description{
Print the primary elements of the "\code{abess}" object.
}
\details{
prints the fitted model and returns it invisibly.
}
\seealso{
\code{\link{abess}}, \code{\link{coef.abess}},
}
