% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.abessrpca.R
\name{print.abessrpca}
\alias{print.abessrpca}
\title{Print method for a fitted "\code{abessrpca}" object}
\usage{
\method{print}{abessrpca}(x, digits = max(5, getOption("digits") - 5), ...)
}
\arguments{
\item{x}{A "\code{abessrpca}" object.}

\item{digits}{Minimum number of significant digits to be used.}

\item{...}{additional print arguments}
}
\value{
No return value, called for side effects
}
\description{
Prints the fitted model and returns it invisibly.
}
\details{
Print a \code{data.frame} with three columns:
the first column is support size of model;
the second column is the explained variance of model;
the last column is the percent of explained variance of model.
}
