% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abess-package.R
\docType{package}
\name{abess-package}
\alias{abess}
\alias{abess-package}
\title{abess: abess: Adaptive Best Subset Selection in Polynomial Time}
\description{
More about what it does (maybe more than one line)
}
\keyword{internal}
