% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abess-package.R
\docType{package}
\name{abess-package}
\alias{abess-package}
\alias{_PACKAGE}
\title{abess: Adaptive Best Subset Selection in Polynomial Time}
\description{
A polynomial algorithm for best-subset selection problem for linear and logistic regression.
}
\author{
\strong{Maintainer}: Jin Zhu \email{zhuj37@mail2.sysu.edu.cn}

Authors:
\itemize{
  \item Kangkang Jiang \email{jiangkk3@mail2.sysu.edu.cn}
  \item Junxian Zhu \email{zhujunx@mail2.sysu.edu.cn}
  \item Canhong Wen \email{wencanhong@gmail.com}
  \item Heping Zhang \email{heping.zhang@yale.edu}
  \item Xueqin Wang \email{wangxq20@ustc.edu.cn}
}

}
\keyword{internal}
