% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abess-package.R
\docType{package}
\name{abess-package}
\alias{abess-package}
\alias{_PACKAGE}
\title{abess: Adaptive Best Subset Selection in Polynomial Time}
\description{
A polynomial algorithm for best-subset selection problem for linear and logistic regression.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/abess-team/abess}
  \item Report bugs at \url{https://github.com/abess-team/abess/issues}
}

}
\author{
\strong{Maintainer}: Jin Zhu \email{zhuj37@mail2.sysu.edu.cn} (\href{https://orcid.org/0000-0001-8550-5822}{ORCID})

Authors:
\itemize{
  \item Kangkang Jiang \email{jiangkk3@mail2.sysu.edu.cn}
  \item Junxian Zhu \email{zhujunx@mail2.sysu.edu.cn}
  \item Canhong Wen \email{wencanhong@gmail.com}
  \item Heping Zhang \email{heping.zhang@yale.edu} (\href{https://orcid.org/0000-0002-0688-4076}{ORCID})
  \item Xueqin Wang \email{wangxq20@ustc.edu.cn} (\href{https://orcid.org/0000-0001-5205-9950}{ORCID})
}

}
\keyword{internal}
