\name{abess-package}
\alias{abess-package}
\alias{abess}
\docType{package}
\title{
\packageTitle{abess}
}
\description{
\packageDescription{abess}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{abess}
\packageIndices{abess}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{abess}

Maintainer: \packageMaintainer{abess}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
