% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abess-package.R
\docType{package}
\name{abess-package}
\alias{abess-package}
\alias{_PACKAGE}
\title{abess: Adaptive Best Subset Selection in Polynomial Time}
\description{
Extremely efficient toolkit for solving the best subset selection problem in linear regression, logistic regression, Poisson regression, Cox proportional hazard model, multiple-response regression, multinomial logistic regression, and (sequential) principal component analysis. It implements and generalizes algorithms designed in <doi:10.1073/pnas.2014241117> that exploits a novel sequencing-and-splicing technique to guarantee exact support recovery and globally optimal solution in polynomial times.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/abess-team/abess}
  \item \url{https://abess-team.github.io/abess/}
  \item \url{https://abess.readthedocs.io}
  \item Report bugs at \url{https://github.com/abess-team/abess/issues}
}

}
\author{
\strong{Maintainer}: Jin Zhu \email{zhuj37@mail2.sysu.edu.cn} (\href{https://orcid.org/0000-0001-8550-5822}{ORCID})

Authors:
\itemize{
  \item Kangkang Jiang \email{jiangkk3@mail2.sysu.edu.cn}
  \item Yanhang Zhang \email{zhangyh98@ruc.edu.cn}
  \item Liyuan Hu \email{huly5@mail2.sysu.edu.cn}
  \item Junhao Huang \email{huangjh256@mail2.sysu.edu.can}
  \item Junxian Zhu \email{zhujunx@mail2.sysu.edu.cn}
  \item Canhong Wen \email{wencanhong@gmail.com}
  \item Heping Zhang \email{heping.zhang@yale.edu} (\href{https://orcid.org/0000-0002-0688-4076}{ORCID})
  \item Xueqin Wang \email{wangxq20@ustc.edu.cn} (\href{https://orcid.org/0000-0001-5205-9950}{ORCID})
}

Other contributors:
\itemize{
  \item spectra contributors (Spectra implementation) [copyright holder]
}

}
\keyword{internal}
