% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abess-package.R
\docType{package}
\name{abess-package}
\alias{abess-package}
\alias{_PACKAGE}
\title{abess: Fast Best Subset Selection}
\description{
Extremely efficient toolkit for solving the best subset selection problem <arXiv:2110.09697>. This package is its R interface. The package implements and generalizes algorithms designed in <doi:10.1073/pnas.2014241117> that exploits a novel sequencing-and-splicing technique to guarantee exact support recovery and globally optimal solution in polynomial times for linear model. It also supports best subset selection for logistic regression, Poisson regression, Cox proportional hazard model, Gamma regression, multiple-response regression, multinomial logistic regression, (sequential) principal component analysis, and robust principal component analysis. The other valuable features such as the best subset of group selection <arXiv:2104.12576> and sure independence screening <doi:10.1111/j.1467-9868.2008.00674.x> are also provided.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/abess-team/abess}
  \item \url{https://abess-team.github.io/abess/}
  \item \url{https://abess.readthedocs.io}
  \item Report bugs at \url{https://github.com/abess-team/abess/issues}
}

}
\author{
\strong{Maintainer}: Jin Zhu \email{zhuj37@mail2.sysu.edu.cn} (\href{https://orcid.org/0000-0001-8550-5822}{ORCID})

Authors:
\itemize{
  \item Liyuan Hu \email{huly5@mail2.sysu.edu.cn}
  \item Junhao Huang \email{huangjh256@mail2.sysu.edu.cn}
  \item Kangkang Jiang \email{jiangkk3@mail2.sysu.edu.cn}
  \item Yanhang Zhang \email{zhangyh98@ruc.edu.cn}
  \item Zezhi Wang \email{homura@mail.ustc.edu.cn}
  \item Borui Tang \email{tangborui@mail.ustc.edu.cn}
  \item Shiyun Lin \email{shiyunlin@stu.pku.edu.cn}
  \item Junxian Zhu \email{adaizjx@163.com}
  \item Canhong Wen \email{wencanhong@gmail.com}
  \item Heping Zhang \email{heping.zhang@yale.edu} (\href{https://orcid.org/0000-0002-0688-4076}{ORCID})
  \item Xueqin Wang \email{wangxq20@ustc.edu.cn} (\href{https://orcid.org/0000-0001-5205-9950}{ORCID})
}

Other contributors:
\itemize{
  \item spectra contributors (Spectra implementation) [copyright holder]
}

}
\keyword{internal}
