% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.abess.R
\name{deviance.abess}
\alias{deviance.abess}
\title{Extract the deviance from a fitted "\code{abess}" object.}
\usage{
\method{deviance}{abess}(object, type = c("standard", "gic", "ebic", "bic", "aic"), ...)
}
\arguments{
\item{object}{A "\code{abess}" object.}

\item{type}{The type of deviance.
One of the following: \code{"standard"},
\code{"gic"}, \code{"ebic"}, \code{"bic"} and \code{"aic"}.
Default is \code{"standard"}.}

\item{...}{additional arguments}
}
\value{
A numeric vector.
}
\description{
Similar to other deviance methods,
which returns deviance from a fitted "\code{abess}" object.
}
\seealso{
\code{\link{print.abess}},
\code{\link{predict.abess}},
\code{\link{coef.abess}},
\code{\link{extract.abess}},
\code{\link{plot.abess}},
\code{\link{deviance.abess}}.
}
