% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.graph.R
\name{generate.bmn.data}
\alias{generate.bmn.data}
\title{Generate a dataset with independent observations from binary markov network}
\usage{
generate.bmn.data(
  n,
  p,
  type = 1,
  seed = NULL,
  graph.seed = NULL,
  beta = 0.7,
  degree = 3,
  alpha = 0.4
)
}
\arguments{
\item{n}{}

\item{p}{}

\item{type}{}

\item{seed}{}
}
\value{

}
\description{
Generate a dataset with independent observations from binary markov network
}
