% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.graph.R
\name{generate.bmn.data}
\alias{generate.bmn.data}
\title{Generate a dataset from binary markov network}
\usage{
generate.bmn.data(
  n,
  p,
  type = 1,
  seed = NULL,
  graph.seed = NULL,
  beta = 0.7,
  degree = 3,
  alpha = 0.4,
  method = "freq"
)
}
\arguments{
\item{n}{The number of observations.}

\item{p}{The number of predictors of interest.}

\item{seed}{random seed. Default: \code{seed = 1}.}

\item{beta}{The coefficient values in the underlying regression model.
If it is supplied, \code{support.size} would be omitted.}

\item{alpha}{}
}
\value{

}
\description{
Generate a dataset from binary markov network
}
