% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.abessrpca.R
\name{plot.abessrpca}
\alias{plot.abessrpca}
\title{Creat plot from a fitted "\code{abessrpca}" object}
\usage{
\method{plot}{abessrpca}(x, type = c("S", "loss", "tune"), support.size = NULL, label = TRUE, ...)
}
\arguments{
\item{x}{A "\code{abessrpca}" object.}

\item{type}{The plot type.
One of the following:
\code{"S"} (i.e., a heatmap for the sparse matrix estimation),
\code{"loss"} (i.e., a support.size versus loss plot),
and \code{"tune"} (i.e., , a support.size versus tuning value plot).
Default is \code{"coef"}.}

\item{support.size}{An integer vector specifies
the sparse matrix fitted at given \code{support.size} to be returned.
If \code{support.size = NULL}, then the sparse matrix with 
the least tuning value would be returned.
Default: \code{support.size = NULL}.}

\item{label}{A logical value.
If \code{label = TRUE} (the default),
label the curves with variable sequence numbers.}

\item{...}{Other graphical parameters to \code{plot} 
or \code{stats::heatmap} function}
}
\value{
No return value, called for side effects.
}
\description{
Produces a sparse-matrix/loss/tuning-value plot
for a fitted "\code{abessrpca}" object.
}
