% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.abess.R
\name{print.abess}
\alias{print.abess}
\title{Print method for a fitted "\code{abess}" object}
\usage{
\method{print}{abess}(x, digits = max(5, getOption("digits") - 5), ...)
}
\arguments{
\item{x}{A "\code{abess}" object.}

\item{digits}{Minimum number of significant digits to be used.}

\item{...}{additional print arguments}
}
\value{
No return value, called for side effects
}
\description{
Prints the fitted model and returns it invisibly.
}
\details{
Print a \code{data.frame} with three columns:
the first column is support size of model;
the second column is deviance of model;
the last column is the tuning value of the certain tuning type.
}
\seealso{
\code{\link{print.abess}},
\code{\link{predict.abess}},
\code{\link{coef.abess}},
\code{\link{extract.abess}},
\code{\link{plot.abess}},
\code{\link{deviance.abess}}.
}
